Add deleteProposalNonAuthor test

+++ src/main/java/eu/trustdemocracy/proposals/core/interactors/proposal/DeleteProposal.java
@@ -3,6 +3,8 @@ package eu.trustdemocracy.proposals.core.interactors.proposal;
 import eu.trustdemocracy.proposals.core.entities.util.ProposalMapper;
 import eu.trustdemocracy.proposals.core.entities.util.UserMapper;
 import eu.trustdemocracy.proposals.core.interactors.Interactor;
+import eu.trustdemocracy.proposals.core.interactors.exceptions.NotAllowedActionException;
+import eu.trustdemocracy.proposals.core.interactors.exceptions.ResourceNotFoundException;
 import eu.trustdemocracy.proposals.core.models.request.ProposalRequestDTO;
 import eu.trustdemocracy.proposals.core.models.response.ProposalResponseDTO;
 import eu.trustdemocracy.proposals.gateways.ProposalDAO;
@@ -18,6 +20,20 @@ public class DeleteProposal implements Interactor<ProposalRequestDTO, ProposalRe
 
   public ProposalResponseDTO execute(ProposalRequestDTO inputProposal) {
     val user = UserMapper.createEntity(inputProposal.getAuthorToken());
+
+    val foundProposal = proposalDAO.findById(inputProposal.getId());
+
+    if (foundProposal == null) {
+      throw new ResourceNotFoundException(
+          "Trying to delete non-existing proposal [" + inputProposal.getId() + "]");
+    }
+
+    if (!foundProposal.getAuthor().getId().equals(user.getId())) {
+      throw new NotAllowedActionException(
+          "User [" + user.getId() + "] is not the owner of proposal ["
+              + foundProposal.getId() + "]");
+    }
+
     val proposal = proposalDAO.delete(inputProposal.getId());
     return ProposalMapper.createResponse(proposal);
   }
